requirements.txt
absl-py==0.6.1
astor==0.7.1
backports.weakref==1.0.post1
bleach==1.5.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
grpcio==1.16.1
html5lib==0.9999999
Markdown==3.0.1
mock==2.0.0
numpy==1.15.4
paho-mqtt==1.4.0
pbr==5.1.1
protobuf==3.6.1
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
Werkzeug==0.14.1
